将净值额度代码重构到一个函数中

+++ src/main/java/com/gofobao/framework/lend/vo/response/LendInfo.java
@@ -43,7 +43,7 @@ public class LendInfo {
     private String equityLimit;
 
     @ApiModelProperty("净值额度隐藏")
-    private Integer equityLimitHide;
+    private Long equityLimitHide;
 
     @ApiModelProperty("状态（0、可借；1、结束）")
     private Integer status;
